requirements.txt
Flask==2.1.0
Pillow==8.4.0
opencv-python-headless==4.5.3.56
pytesseract==0.3.8